printf: remove cast
Since no warning was reported by gcc, so remove cast

+++ printf.c
@@ -6,7 +6,7 @@ int main()
 	char s[] = "uvc";
 
 	printf("%i\n", a);
-	printf("%.*s\n", (int)sizeof(s), s);
+	printf("%.*s\n", sizeof(s), s);
 
 	printf("This is a printf test case, join "
 			"two lines together\n");
